+++ src/Stheno.jl
@@ -36,33 +36,24 @@ module Stheno
     include(joinpath("util", "zygote_rules.jl"))
     include(joinpath("util", "covariance_matrices.jl"))
     include(joinpath("util", "block_arrays.jl"))
-    include(joinpath("util", "abstract_data_set.jl"))
-    include(joinpath("util", "proper_type_piracy.jl"))
 
-    # Supertype for GPs.
-    include("abstract_gp.jl")
-
-    # Atomic GP objects.
-    include(joinpath("gp", "gp.jl"))
+    # New AbstractGPs and associated utility.
+    include(joinpath("gp", "util.jl"))
+    include(joinpath("gp", "atomic_gp.jl"))
+    include(joinpath("gp", "derived_gp.jl"))
+    include(joinpath("gp", "sparse_finite_gp.jl"))
 
     # Composite GPs, constructed via affine transformation of CompositeGPs and GPs.
-    include(joinpath("composite", "composite_gp.jl"))
-    include(joinpath("composite", "cross.jl"))
-    include(joinpath("composite", "product.jl"))
-    include(joinpath("composite", "addition.jl"))
-    include(joinpath("composite", "compose.jl"))
-    # include(joinpath("composite", "gradient.jl"))
-    # include(joinpath("composite", "integrate.jl"))
+    include(joinpath("affine_transformations", "cross.jl"))
+    include(joinpath("affine_transformations", "product.jl"))
+    include(joinpath("affine_transformations", "addition.jl"))
+    include(joinpath("affine_transformations", "compose.jl"))
+    # include(joinpath("affine_transformations", "gradient.jl"))
+    # include(joinpath("affine_transformations", "integrate.jl"))
 
-    # Gaussian Process Probabilistic Programme object which implements the AbstractGPs API.
     include("gaussian_process_probabilistic_programme.jl")
 
-    # Sparse GP hack to make pseudo-point approximations play nicely with Turing.jl.
-    include("sparse_finite_gp.jl")
-
-    include("deprecate.jl")
-
-    export wrap, BlockData, GPC, GPPPInput, @gppp
+    export atomic, BlockData, GPC, GPPPInput, @gppp
     export ∘, select, stretch, periodic, shift
     export cov_diag, mean_and_cov_diag
 end # module