+++ src/deprecate.jl
@@ -1,7 +1,0 @@
-@deprecate GP(m, k, gpc) Stheno.wrap(GP(m, k), gpc)
-@deprecate GP(k, gpc) Stheno.wrap(GP(k), gpc)
-
-using KernelFunctions: NeuralKernelNetwork
-export NeuralKernelNetwork
-
-@deprecate approx_posterior posterior