+++ src/gp/derived_gp.jl
@@ -1,28 +1,28 @@
 """
-    CompositeGP{Targs} <: AbstractGP
+    DerivedGP{Targs} <: AbstractGP
 
 A GP derived from other GPs via an affine transformation. Specification given by `args`.
 You should generally _not_ construct this object manually.
 """
-struct CompositeGP{Targs} <: SthenoAbstractGP
+struct DerivedGP{Targs} <: SthenoAbstractGP
     args::Targs
     n::Int
     gpc::GPC
-    function CompositeGP{Targs}(args::Targs, gpc::GPC) where {Targs}
+    function DerivedGP{Targs}(args::Targs, gpc::GPC) where {Targs}
         gp = new{Targs}(args, next_index(gpc), gpc)
         gpc.n += 1
         return gp
     end
 end
-CompositeGP(args::Targs, gpc::GPC) where {Targs} = CompositeGP{Targs}(args, gpc)
+DerivedGP(args::Targs, gpc::GPC) where {Targs} = DerivedGP{Targs}(args, gpc)
 
-mean(f::CompositeGP, x::AbstractVector) = mean(f.args, x)
+mean(f::DerivedGP, x::AbstractVector) = mean(f.args, x)
 
-cov(f::CompositeGP, x::AbstractVector) = cov(f.args, x)
-var(f::CompositeGP, x::AbstractVector) = var(f.args, x)
+cov(f::DerivedGP, x::AbstractVector) = cov(f.args, x)
+var(f::DerivedGP, x::AbstractVector) = var(f.args, x)
 
-cov(f::CompositeGP, x::AbstractVector, x′::AbstractVector) = cov(f.args, x, x′)
-var(f::CompositeGP, x::AbstractVector, x′::AbstractVector) = var(f.args, x, x′)
+cov(f::DerivedGP, x::AbstractVector, x′::AbstractVector) = cov(f.args, x, x′)
+var(f::DerivedGP, x::AbstractVector, x′::AbstractVector) = var(f.args, x, x′)
 
 function cov(
     f::SthenoAbstractGP, f′::SthenoAbstractGP, x::AbstractVector, x′::AbstractVector,
@@ -30,7 +30,7 @@ function cov(
     @assert f.gpc === f′.gpc
     if f.n === f′.n
         return cov(f.args, x, x′)
-    elseif f isa WrappedGP && f.n > f′.n || f′ isa WrappedGP && f′.n > f.n
+    elseif f isa AtomicGP && f.n > f′.n || f′ isa AtomicGP && f′.n > f.n
         return zeros(length(x), length(x′))
     elseif f.n >= f′.n
         return cov(f.args, f′, x, x′)
@@ -45,7 +45,7 @@ function var(
     @assert f.gpc === f′.gpc
     if f.n === f′.n
         return var(f.args, x, x′)
-    elseif f isa WrappedGP && f.n > f′.n || f′ isa WrappedGP && f′.n > f.n
+    elseif f isa AtomicGP && f.n > f′.n || f′ isa AtomicGP && f′.n > f.n
         return zeros(length(x))
     elseif f.n >= f′.n
         return var(f.args, f′, x, x′)