+++ src/util/proper_type_piracy.jl
@@ -1,3 +1,0 @@
-# This file contains some real first-rate type piracy. Sorry in advance.
-
-LinearAlgebra.UpperTriangular(D::Diagonal) = D