+++ test/affine_transformations/cross.jl
@@ -2,8 +2,8 @@
     @timedtestset "Correctness tests" begin
         rng, P, Q, gpc = MersenneTwister(123456), 2, 3, GPC()
 
-        f1 = wrap(GP(sin, SEKernel()), gpc)
-        f2 = wrap(GP(cos, SEKernel()), gpc)
+        f1 = atomic(GP(sin, SEKernel()), gpc)
+        f2 = atomic(GP(cos, SEKernel()), gpc)
         f3 = cross([f1, f2])
         f4 = cross([f1])
         f5 = cross([f2])
@@ -51,8 +51,8 @@
         x2, x3 = randn(rng, P), randn(2P)
 
         gpc = GPC()
-        f1 = wrap(GP(sin, SEKernel()), gpc)
-        f2 = wrap(GP(cos, SEKernel()), gpc)
+        f1 = atomic(GP(sin, SEKernel()), gpc)
+        f2 = atomic(GP(cos, SEKernel()), gpc)
         f3 = cross([f1, f2])
         abstractgp_interface_tests(f3, f1, x0, x1, x2, x3)
 
@@ -76,8 +76,8 @@
         rng, P, Q = MersenneTwister(123456), 3, 5
         xp, xq = randn(rng, P), randn(rng, Q)
         gpc = GPC()
-        f_ = wrap(GP(sin, SEKernel()), gpc)
-        g_ = wrap(GP(cos, SEKernel()), gpc)
+        f_ = atomic(GP(sin, SEKernel()), gpc)
+        g_ = atomic(GP(cos, SEKernel()), gpc)
         h_x = Stheno.finites_to_block([f_(xp, 1.0), g_(xq, 1.0)])
     end
 end