+++ test/gaussian_process_probabilistic_programme.jl
@@ -1,5 +1,45 @@
 @timedtestset "gaussian_process_probabilistic_programme" begin
 
+    @timedtestset "BlockData" begin
+
+        rng = MersenneTwister(123456)
+        N = 10
+        D = 2
+        x = randn(rng, N)
+        X = randn(rng, D, N)
+
+        DX = ColVecs(X)
+        DxX = BlockData([x, DX])
+        @test size(DxX) == (2N,)
+        @test length(DxX) == 2N
+        @test DxX == DxX
+        @test DxX == BlockData([x, DX])
+        @test length([x for x in DxX]) == 2N
+        @test getindex(DxX, 1) === x[1]
+        @test getindex(DxX, 2) === x[2]
+        @test getindex(DxX, N) === x[N]
+        @test getindex(DxX, N + 1) == DX[1]
+        @test view(DxX, 2, 1) == view(DX, 1)
+        @test view(DxX, 2, N) == view(DX, N)
+        @test eachindex(DxX) isa BlockVector
+        @test eachindex(DxX) == mortar([1:N, N+1:2N], [N, N])
+
+        # Test iteration.
+        @test [x for x in DxX][1] == x[1]
+        @test [x for x in DxX][N] == x[end]
+        @test [x for x in DxX][N + 1] == DX[1]
+        @test [x for x in DxX][end] == DX[end]
+
+        # Test that identically typed arrays yield the appropriate element type.
+        @test eltype(BlockData([randn(5), randn(4)])) == Float64
+        @test eltype(DxX) == Any
+        @test eltype(BlockData([DX, DX])) <: AbstractVector{Float64}
+        @test eltype(BlockData([eachindex(DX), eachindex(DX)])) == Int
+
+        # Convenience constructor.
+        @test BlockData(x, DX) == DxX
+    end
+
     @timedtestset "split" begin
         x = BlockData(randn(5), randn(4))
         @testset "Vector" begin
@@ -16,8 +56,8 @@
 
     # Build a toy collection of processes.
     gpc = GPC()
-    f1 = wrap(GP(sin, SEKernel()), gpc)
-    f2 = wrap(GP(cos, Matern52Kernel()), gpc)
+    f1 = atomic(GP(sin, SEKernel()), gpc)
+    f2 = atomic(GP(cos, Matern52Kernel()), gpc)
     f3 = f1 + 3 * f2
 
     # Use them to build a programme.
@@ -107,7 +147,7 @@
     @timedtestset "nested gppp" begin
 
         gpc_outer = GPC()
-        f1_outer = Stheno.wrap(f, gpc_outer)
+        f1_outer = Stheno.atomic(f, gpc_outer)
         f2_outer = 5 * f1_outer
         f_outer = Stheno.GPPP((f1=f1_outer, f2=f2_outer), gpc_outer)