+++ test/gp/derived_gp.jl
@@ -1,0 +1,5 @@
+@timedtestset "derived_gp" begin
+    
+end
+
+# This type is tested in affine_transformations