+++ test/gp/sparse_finite_gp.jl
@@ -3,18 +3,18 @@
     xu = 0:10
     σ = 1.0
     σu = 1e-3
-    f = wrap(GP(Matern32Kernel()), GPC())
+    f = atomic(GP(Matern32Kernel()), GPC())
     covariance_error = "The covariance matrix of a sparse GP can often be dense and " *
         "can cause the computer to run out of memory. If you are sure you have enough " *
         "memory, you can use `cov(f.fobs)`."
 
     @timedtestset "SparseFiniteGP Constructors" begin
-        f = wrap(GP(Matern32Kernel()), GPC())
+        f = atomic(GP(Matern32Kernel()), GPC())
         @test SparseFiniteGP(f(x), f(xu)) == SparseFiniteGP(f(x, 1e-18), f(xu, 1e-18))
     end
 
     @timedtestset "SparseFiniteGP methods" begin
-        f = wrap(GP(Matern32Kernel()), GPC())
+        f = atomic(GP(Matern32Kernel()), GPC())
         fx = f(x)
         fxu = SparseFiniteGP(f(x), f(xu))
         @test mean(fxu) == mean(fx)
@@ -25,7 +25,7 @@
     end
 
     @timedtestset "SparseFiniteGP inference" begin
-        f = wrap(GP(Matern32Kernel()), GPC())
+        f = atomic(GP(Matern32Kernel()), GPC())
         fx = f(x, σ)
         fxu = SparseFiniteGP(f(x, σ), f(xu, σu))
         y = rand(MersenneTwister(12345), fxu)