+++ test/gp/util.jl
@@ -9,7 +9,7 @@ end
     @testset "statistics" begin
         rng, N, N′ = MersenneTwister(123456), 1, 9
         x, x′, Σy, Σy′ = randn(rng, N), randn(rng, N′), zeros(N, N), zeros(N′, N′)
-        f = wrap(GP(sin, SEKernel()), GPC())
+        f = atomic(GP(sin, SEKernel()), GPC())
         fx, fx′ = FiniteGP(f, x, Σy), FiniteGP(f, x′, Σy′)
 
         @test mean(fx) == mean(f, x)
@@ -23,8 +23,8 @@ end
         rng, N, D = MersenneTwister(123456), 10, 2
         X, x, Σy = ColVecs(randn(rng, D, N)), randn(rng, N), zeros(N, N)
         Σy = generate_noise_matrix(rng, N)
-        fX = FiniteGP(wrap(GP(1, SEKernel()), GPC()), X, Σy)
-        fx = FiniteGP(wrap(GP(1, SEKernel()), GPC()), x, Σy)
+        fX = FiniteGP(atomic(GP(1, SEKernel()), GPC()), X, Σy)
+        fx = FiniteGP(atomic(GP(1, SEKernel()), GPC()), x, Σy)
 
         # Check that single-GP samples have the correct dimensions.
         @test length(rand(rng, fX)) == length(X)
@@ -36,7 +36,7 @@ end
     @testset "rand (statistical)" begin
         rng, N, D, μ0, S = MersenneTwister(123456), 10, 2, 1, 100_000
         X, Σy = ColVecs(randn(rng, D, N)), 1e-12
-        f = FiniteGP(wrap(GP(1, SEKernel()), GPC()), X, Σy)
+        f = FiniteGP(atomic(GP(1, SEKernel()), GPC()), X, Σy)
 
         # Check mean + covariance estimates approximately converge for single-GP sampling.
         f̂ = rand(rng, f, S)
@@ -54,7 +54,7 @@ end
         adjoint_test(
             x->rand(
                 MersenneTwister(123456),
-                FiniteGP(wrap(GP(sin, SEKernel()), GPC()), x, Σy),
+                FiniteGP(atomic(GP(sin, SEKernel()), GPC()), x, Σy),
             ),
             randn(rng, N),
             x;
@@ -65,7 +65,7 @@ end
         adjoint_test(
             x->rand(
                 MersenneTwister(123456),
-                FiniteGP(wrap(GP(sin, SEKernel()), GPC()), x, Σy),
+                FiniteGP(atomic(GP(sin, SEKernel()), GPC()), x, Σy),
                 S,
             ),
             randn(rng, N, S),
@@ -77,7 +77,7 @@ end
         rng = MersenneTwister(123456)
         x = randn(rng, T, 123)
         z = randn(rng, T, 13)
-        f = wrap(GP(T(0), SEKernel()), GPC())
+        f = atomic(GP(T(0), SEKernel()), GPC())
 
         fx = f(x, T(0.1))
         u = f(z, T(1e-4))