+++ test/runtests.jl
@@ -55,33 +55,26 @@ include("test_util.jl")
         include(joinpath("util", "zygote_rules.jl"))
         include(joinpath("util", "covariance_matrices.jl"))
         include(joinpath("util", "block_arrays.jl"))
-        include(joinpath("util", "abstract_data_set.jl"))
     end
 
     println("gp:")
     @timedtestset "gp" begin
-        include(joinpath("gp", "gp.jl"))
+        include(joinpath("gp", "util.jl"))
+        include(joinpath("gp", "atomic_gp.jl"))
+        include(joinpath("gp", "derived_gp.jl"))
+        include(joinpath("gp", "sparse_finite_gp.jl"))
     end
 
-    println("composite:")
-    @timedtestset "composite" begin
-        include(joinpath("composite", "test_util.jl"))
-        include(joinpath("composite", "cross.jl"))
-        include(joinpath("composite", "product.jl"))
-        include(joinpath("composite", "addition.jl"))
-        include(joinpath("composite", "compose.jl"))
+    println("affine_transformations:")
+    @timedtestset "affine_transformations" begin
+        include(joinpath("affine_transformations", "test_util.jl"))
+        include(joinpath("affine_transformations", "addition.jl"))
+        include(joinpath("affine_transformations", "compose.jl"))
+        include(joinpath("affine_transformations", "product.jl"))
     end
 
-    println("abstract_gp:")
-    @timedtestset "abstract_gp" begin
-        include("abstract_gp.jl")
-    end
-
-    println("gaussian_process_probabilistic_programme:")
+    include(joinpath("affine_transformations", "cross.jl"))
     include("gaussian_process_probabilistic_programme.jl")
-
-    println("sparse_finite_gp:")
-    include("sparse_finite_gp.jl")
 
     println("doctests")
     @timedtestset "doctests" begin